# [WIP] free speed/mem optimizations with ahash, dary_heap, and compact_str