--- a/tokenizers/src/models/mod.rs
+++ b/tokenizers/src/models/mod.rs
@@ -5,6 +5,7 @@ pub mod unigram;
 pub mod wordlevel;
 pub mod wordpiece;
 
+use ahash::AHashMap;
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
@@ -19,11 +20,11 @@ use crate::{AddedToken, Model, Result, Token, Trainer};
 /// Wraps a vocab mapping (ID -> token) to a struct that will be serialized in order
 /// of token ID, smallest to largest.
 struct OrderedVocabIter<'a> {
-    vocab_r: &'a HashMap<u32, String>,
+    vocab_r: &'a AHashMap<u32, String>,
 }
 
 impl<'a> OrderedVocabIter<'a> {
-    fn new(vocab_r: &'a HashMap<u32, String>) -> Self {
+    fn new(vocab_r: &'a AHashMap<u32, String>) -> Self {
         Self { vocab_r }
     }
 }
@@ -179,6 +180,15 @@ impl Model for ModelWrapper {
         }
     }
 
+    fn get_vocab_ahash(&self) -> AHashMap<String, u32> {
+        match self {
+            Self::WordLevel(t) => t.get_vocab_ahash(),
+            Self::WordPiece(t) => t.get_vocab_ahash(),
+            Self::BPE(t) => t.get_vocab_ahash(),
+            Self::Unigram(t) => t.get_vocab_ahash(),
+        }
+    }
+
     fn get_vocab_size(&self) -> usize {
         match self {
             Self::WordLevel(t) => t.get_vocab_size(),
@@ -284,8 +294,8 @@ mod tests {
 
     #[test]
     fn incomplete_ordered_vocab() {
-        let vocab_r: HashMap<u32, String> =
-            HashMap::from([(0, "Hi".to_string()), (2, "There".to_string())]);
+        let vocab_r: AHashMap<u32, String> =
+            AHashMap::from([(0, "Hi".to_string()), (2, "There".to_string())]);
 
         let ordered = OrderedVocabIter::new(&vocab_r);