@McPatate 2025-03-24
> (although there was already ~420K leaked with cargo bench on HEAD)

🥲 

Super cool work! Would be cool to see which of `FxHashMap` & `AHash` performs best.

In the meantime, feel free to open a smaller PR with only the `dary_heap` addition, which afaiu doesn't overlap with @MeetThePatel's work