@mjbommar 2025-03-24
@McPatate, sure, I can repackage this - although @MeetThePatel is also welcome to cherry-pick or take inspiration from this too.

PS: As far as FxHash vs AHash goes, I think we might even get even better performanec thanks to recent optimality breakthough:

- https://arxiv.org/abs/2501.02305
- https://github.com/sternma/optopenhash
- https://crates.io/crates/elastic_hash_rs

Ideally one of the existing, mature hash libraries would pick up the key ideas, but that doens't seem to have happened yet.