@MeetThePatel 2025-03-24
From my understanding, the hashmaps presented in the paper are optimized for reducing worst-case performance. I quickly glanced at the algorithm a couple of weeks ago, and I suspect the wall-clock performance of average case would be lower than SwissTable (what std::collections::HashMap uses), as SwissTable has been SIMD optimized, although I haven't tested this at all, just a hunch.